feat: move knowledge DTO contracts to shared crate
