+++ app/src-tauri/Cargo.toml
@@ -34,6 +34,7 @@ time = "=0.3.51"
 
 # Lõi convert dùng chung với CLI — KHÔNG viết lại.
 fileconv-core = { path = "../../crates/core", features = ["llm"] }
+fileconv-knowledge = { path = "../../crates/knowledge" }
 rusqlite = { version = "0.37.0", features = ["bundled"], default-features = false }
 notify = "8.2.0"
 hnsw_rs = "0.3.4"
