+++ plans/markhand-web/roadmap.html
@@ -919,7 +919,7 @@
     };
 
     /* ROADMAP_DATA_START */
-    const ROADMAP_SOURCE_HASH = "3b88d9be8e93f9b8";
+    const ROADMAP_SOURCE_HASH = "b4acca0359dbcec2";
     const phases = [
       {
         "id": "phase-f",
@@ -1072,7 +1072,7 @@
           [
             "P1A-03",
             "Shared DTO và serde contract",
-            "ready"
+            "in_progress"
           ],
           [
             "P1A-04",
